% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimps.R
\name{graph_results}
\alias{graph_results}
\title{graph_results}
\usage{
graph_results(results, object)
}
\arguments{
\item{results}{The results from calc_vimps}

\item{object}{Which object from results to use for graphing results}
}
\value{
No return value
}
\description{
Graph the variable importance results from calc_vimps
}
